backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
django-import-export==4.3.3
Pillow==11.3.0
pandas==2.3.2
numpy==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
gunicorn==23.0.0
whitenoise==6.8.2
psycopg2-binary==2.9.10